Handle monotonic time

+++ config.go
@@ -94,28 +94,32 @@ func migrateConfig() {
 	setConfigValue("cli", "config-version", configVersion)
 	
 	cliPath, _ := getAkamaiCliPath()
-	upgradeFile := filepath.Join(cliPath, ".upgrade-check")
 
 	setConfigValue("cli", "last-upgrade-check", "never")
 
+	var data []byte
+	upgradeFile := filepath.Join(cliPath, ".upgrade-check")
 	if _, err := os.Stat(upgradeFile); err == nil {
-		data, err := ioutil.ReadFile(upgradeFile)
-		lastUpgrade, err := time.Parse("2006-01-02 15:04:05.999999999 -0700 MST", string(data))
-		if err == nil {
-			setConfigValue("cli", "last-upgrade-check", lastUpgrade.Format(time.RFC3339))
-		}
+		data, err = ioutil.ReadFile(upgradeFile)
 	} else {
 		upgradeFile = filepath.Join(cliPath, ".update-check")
 		if _, err := os.Stat(upgradeFile); err == nil {
-			data, err := ioutil.ReadFile(upgradeFile)
-			lastUpgrade, err := time.Parse("2006-01-02 15:04:05.999999999 -0700 MST", string(data))
-			if err == nil {
-				setConfigValue("cli", "last-upgrade-check", lastUpgrade.Format(time.RFC3339))
-			}
+			data, err = ioutil.ReadFile(upgradeFile)
 		}
 	}
 
-	os.Remove(upgradeFile)
+	if len(data) != 0 {
+		date := string(data)
+		if m := strings.LastIndex(date, "m="); m != -1 {
+			date = date[0:m-1]
+		}
+		lastUpgrade, err := time.Parse("2006-01-02 15:04:05.999999999 -0700 MST", date)
+		if err == nil {
+			setConfigValue("cli", "last-upgrade-check", lastUpgrade.Format(time.RFC3339))
+		}
+
+		os.Remove(upgradeFile)
+	}
 
 	saveConfig()
 }
